Traq.Analysis.CostModel.QueryCost

Documentation

data QueryCost precT

Constructors

QueryCost

Fields

uqueries :: Map Ident precT

cqueries :: Map Ident precT

totalQueryCost

Arguments

:: forall a. Rig a

=> QueryCost a

-> Ticks a
cost per unitary query

-> Ticks a
cost per classical query

-> a

newtype SimpleQueryCost precT

A simple cost that counts the number of queries to all external functions. It treats unitary and classical queries as the same.

Constructors

SimpleQueryCost

Fields

getCost :: precT