Traq.Analysis.Error.Prelude

Failure probability type

data FailProb precT

Wrapper for values representing a failure probability in [0, 1]

failProb :: Num precT => precT -> FailProb precT

Safe constructor

getFailProb :: FailProb precT -> precT

splitFailProb :: Fractional prec => FailProb prec -> prec -> FailProb prec

Helper: split an epsilon into n equal parts.

Conversions

unitarySubroutineTVError :: Floating prec => FailProb prec -> FailProb prec

Given the TV error of a unitary channel, convert it to the contribution as a subroutine via the diamond-distance of the unitary dilations.

unitarySubroutineTVErrorTotal :: Floating prec => prec -> FailProb prec -> FailProb prec

Total error over n calls to subroutine with error eps.

unitarySubroutineTVBudget :: Fractional prec => FailProb prec -> FailProb prec

Given a error budget for subroutine, convert to budget for its unitary dilation.

Monad

type ErrorAnalysisMonad ext = Reader (FunCtx ext)

type ErrorReqs size prec = (Floating prec, Num size, Ord prec, SizeToPrec size prec)