Traq.Compiler.Prelude

Utilities for generating identifiers

type UniqNamesCtx = Set Ident

A set of already used names. This is used to generate new unique identifiers.

class HasUniqNamesCtx s where

Methods

_uniqNamesCtx :: Lens' s UniqNamesCtx

newIdent :: forall s m. (MonadError String m, MonadState s m, HasUniqNamesCtx s) => Ident -> m Ident

Generate a new identifier with the given prefix.

mkQProcName :: Ident -> Ident

get the name of the compiled (cq) proc given the source fun name

mkUProcName :: Ident -> Ident

get the name of the compiled uproc given the source fun name

Compilation Monad

type CompilerT ext = RWST (LoweringEnv ext) (LoweringOutput (SizeType ext)) (LoweringCtx (SizeType ext)) (Either String)

Monad to compile source programs to CQPL programs. This should contain the _final_ typing context for the input program, that is, contains both the inputs and outputs of each statement.

compileWith :: forall ext size m. (m ~ CompilerT ext, size ~ SizeType ext, HasFreeVars ext, TypeInferrable ext size) => (Program ext -> m ()) -> Program ext -> Either String (Program size)

Run the given compiler on a full program.

State

data LoweringCtx sizeT

A global lowering context.

data ProcSignature size

Signature of a compiled uproc/proc: inputs, outputs, ancilla (for unitary)

Constructors

ProcSignature

Fields

in_tys, out_tys, aux_tys :: [VarType size]

_procSignatures :: Lens' (LoweringCtx size) (Map Ident (ProcSignature size))

Output

data LoweringOutput size

The outputs of lowering

_loweredProcs :: Lens' (LoweringOutput size) [ProcDef size]

addProc :: MonadWriter (LoweringOutput size) m => ProcDef size -> m ()

Env

type LoweringEnv ext = FunCtx ext

Read-only compiler env