Traq.Compiler.Quantum

Compilation

lowerProgram

Arguments

:: (Lowerable primsT primsT holeT sizeT costT, TypeCheckable sizeT, Show costT, Floating costT)

=> PrecisionSplittingStrategy

-> TypingCtx sizeT
input bindings to the source program

-> OracleTicks costT
unitary ticks

-> OracleTicks costT
classical ticks

-> costT
fail prob \( \varepsilon \)

-> Program primsT sizeT
source program

-> Either String (Program holeT sizeT costT)

Lower a full program into a CQPL program.

Types

type CompilerT primT holeT sizeT costT = WriterT (LoweringOutput holeT sizeT costT) (ReaderT (LoweringEnv primT sizeT costT) (StateT (LoweringCtx sizeT) (Either String)))

Monad to compile ProtoQB to CQPL programs. This should contain the _final_ typing context for the input program, that is, contains both the inputs and outputs of each statement.

type LoweringEnv primT sizeT costT = QuantumMaxCostEnv primT sizeT costT

Configuration for lowering

data LoweringCtx sizeT

A global lowering context, consisting of - The set of already used identifiers (to generate unique ones) - The typing context: mapping all variables in context to their types.

type LoweringOutput holeT sizeT costT = [ProcDef holeT sizeT costT]

The outputs of lowering

Primitive implementations

class Lowerable primsT primT holeT sizeT costT => Lowerable primsT primT holeT sizeT costT where

Primitives that support a classical-quantum lowering.

Methods

lowerPrimitive

Arguments

:: costT
fail prob

-> primT

-> [Ident]
args

-> [Ident]
rets

-> CompilerT primsT holeT sizeT costT (Stmt holeT sizeT)