Traq.Control.Monad

MonadState

withSandboxOf :: MonadState s m => Lens' s s' -> m a -> m a

Save the current state, run a computation and restore the saved state.

withSandbox :: MonadState s m => m a -> m a

Save the current state, run a computation and restore the saved state.

MonadWriter

tell :: MonadWriter w m => w -> m ()

tell w is an action that produces the output w.

tellAt :: MonadWriter w m => Lens' w w' -> w' -> m ()

Write at a particular location in the monoid.

writeElemAt :: (Applicative f, MonadWriter w m) => Lens' w (f a) -> a -> m ()

writeElem :: (Applicative f, MonadWriter (f a) m) => a -> m ()

MonadError

throwFrom :: MonadError MyError m => m a -> MyError -> m a

try-catch block that prepends a message to the existing error, to produce a more verbose backtrace.

singularM :: MonadError e m => Getting (Endo [a]) s a -> e -> s -> m a

extract a singular value from a traversal, throwing an error if there are 0/more than 1.

maybeWithError :: MonadError e m => e -> Maybe a -> m a

lift a Maybe to a value, throwing an error if Nothing

Helpers

(??) :: Functor f => f (a -> b) -> a -> f b