Traq.Primitives.Class.Compile

Documentation

data PrimCompileEnv shape size

Helpers to compile a primitive.

Constructors

PrimCompileEnv

Fields

mk_ucall :: shape (UCallBuilder size)

helper to generate a call to a unitary function argument.

mk_call :: shape (CallBuilder size)

helper to generate a call to a classical function argument.

mk_meas :: shape (CallBuilder size)

helper to generate a call-and-meas to a unitary proc arg.

uproc_aux_types :: shape [VarType size]

auxiliary variables for each unitary function argument.

ret_vars :: [Ident]

return variables to store the result in.

class (size ~ SizeType prim, prec ~ PrecType prim, ValidPrimShape (PrimFnShape prim)) => UnitaryCompilePrim prim size prec | prim -> size prec where

Compile a primitive to a uproc

Minimal complete definition

Nothing

Methods

compileUPrim :: forall ext' m shape. (m ~ PrimCompileMonad ext' prim, size ~ SizeType ext', prec ~ PrecType ext', shape ~ PrimFnShape prim) => prim -> FailProb prec -> m (ProcDef size)

default compileUPrim :: forall ext' m shape. (Generic prim, GUnitaryCompilePrim (Rep prim) size prec, m ~ PrimCompileMonad ext' prim, size ~ SizeType ext', prec ~ PrecType ext', shape ~ PrimFnShape prim) => prim -> FailProb prec -> m (ProcDef size)

class (size ~ SizeType prim, prec ~ PrecType prim, ValidPrimShape (PrimFnShape prim)) => QuantumCompilePrim prim size prec | prim -> size prec where

Compile a primitive to a cq-proc

Minimal complete definition

Nothing

Methods

compileQPrim :: forall ext' m shape. (m ~ PrimCompileMonad ext' prim, size ~ SizeType ext', prec ~ PrecType ext', shape ~ PrimFnShape prim) => prim -> FailProb prec -> m (ProcDef size)

default compileQPrim :: forall ext' m shape. (Generic prim, GQuantumCompilePrim (Rep prim) size prec, m ~ PrimCompileMonad ext' prim, size ~ SizeType ext', prec ~ PrecType ext', shape ~ PrimFnShape prim) => prim -> FailProb prec -> m (ProcDef size)

type PrimCompileMonad ext prim = ReaderT (PrimCompileEnv (PrimFnShape prim) (SizeType prim)) (CompilerT ext)