Traq.Primitives.Class.Eval

Documentation

class (size ~ SizeType prim, prec ~ PrecType prim, ValidPrimShape (PrimFnShape prim)) => EvalPrim prim size prec | prim -> size prec where

Evaluate a primitive given the semantics of each function argument. For partial functions, the prefix of arguments are already bound.

Minimal complete definition

Nothing

Methods

evalPrim :: forall ext' shape m. (Evaluatable ext' size prec, m ~ Evaluator ext', SizeType ext' ~ size, PrecType ext' ~ prec, shape ~ PrimFnShape prim) => prim -> shape ([Value size] -> m [Value size]) -> m [Value size]

default evalPrim :: forall ext' shape m. (Generic prim, GEvalPrim (Rep prim) size prec, Evaluatable ext' size prec, m ~ Evaluator ext', SizeType ext' ~ size, PrecType ext' ~ prec, shape ~ PrimFnShape prim) => prim -> shape ([Value size] -> m [Value size]) -> m [Value size]