Traq.Primitives.Class.Prelude

Primitives

type family PrimFnShape prim :: Type -> Type

The shape of the function arguments that primitive prim expects. The type PrimFnShape prim a should be a subtype of [a], for every a. This type is useful to specify record constructor names for readability.

class ValidPrimShape shape where

Conversion between shaped information and a list.

Methods

listToShape :: [a] -> Either String (shape a)

shapeToList :: shape a -> [a]

reshape :: (MonadError String m, ValidPrimShape shape, ValidPrimShape shape') => shape a -> m (shape' a)

reshapeUnsafe :: (ValidPrimShape shape, ValidPrimShape shape') => shape a -> shape' a

Partial Functions

data PartialFun

A partially applied function Syntax: f(a_1, ..., a_n) where each a_i is either an identifier, or a blank _

Constructors

PartialFun

Fields

pfun_name :: Ident

pfun_args :: [Maybe Ident]

placeArgs :: [Maybe a] -> [a] -> [a]

Place a list of concrete values inside a list of incomplete values. For example, placeArgs [Just 0, Nothing, Just 2, Nothing] [1, 3] = [0,1,2,3]

placeArgsWithExcess :: [Maybe a] -> [a] -> [a]

Place arguments, with the excess at tail position.