Traq.Primitives.Class.QuantumCost

Classical-Quantum Compilation and Cost

class (size ~ SizeType prim, prec ~ PrecType prim, SizeToPrec size prec, TypeCheckPrim prim size) => QuantumHavocCostPrim prim size prec | prim -> size prec where

Worst-case Quantum query and operation costs of a primitive. Represents one level of the call graph. The quantum compiler of the primitive can use both the quantum and unitary compilation of its function arguments.

Minimal complete definition

Nothing

Methods

quantumQueryCostsQuantum :: prim -> FailProb prec -> PrimFnShape prim prec

Bound on number of queries made to each function's quantum compilation.

default quantumQueryCostsQuantum :: (Generic prim, GQuantumHavocCostPrim (Rep prim) size prec) => prim -> FailProb prec -> PrimFnShape prim prec

quantumQueryCostsUnitary :: prim -> FailProb prec -> PrimFnShape prim (UnitaryQueries prec)

Bound on number of queries made to each function's unitary compilation.

default quantumQueryCostsUnitary :: (Generic prim, GQuantumHavocCostPrim (Rep prim) size prec) => prim -> FailProb prec -> PrimFnShape prim (UnitaryQueries prec)

quantumExprCosts :: (CostModel cost, prec ~ PrecType cost) => prim -> FailProb prec -> cost

Cost of all additional operations.

default quantumExprCosts :: (Generic prim, GQuantumHavocCostPrim (Rep prim) size prec, CostModel cost, prec ~ PrecType cost) => prim -> FailProb prec -> cost

class (size ~ SizeType prim, prec ~ PrecType prim, SizeToPrec size prec, TypeCheckPrim prim size) => QuantumExpCostPrim prim size prec where

Expected Quantum query and operation costs of a primitive. Represents one level of the call graph. The quantum compiler of the primitive can use both the quantum and unitary compilation of its function arguments.

Minimal complete definition

Nothing

Methods

quantumExpQueryCostsQuantum :: forall shape m. (shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> shape [([Value size], prec)]

default quantumExpQueryCostsQuantum :: forall shape m. (Generic prim, GQuantumExpCostPrim (Rep prim) size prec) => (shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> shape [([Value size], prec)]

quantumExpQueryCostsUnitary :: forall shape m. (shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> shape (UnitaryQueries prec)

Bound on the expected number of queries made to each function's unitary compilation.

default quantumExpQueryCostsUnitary :: forall shape m. (Generic prim, GQuantumExpCostPrim (Rep prim) size prec) => (shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> shape (UnitaryQueries prec)

quantumExpExprCosts :: forall shape cost m. (CostModel cost, prec ~ PrecType cost, shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> cost

Cost of all additional operations. Defaults to zero.

default quantumExpExprCosts :: forall shape cost m. (Generic prim, GQuantumExpCostPrim (Rep prim) size prec) => (CostModel cost, prec ~ PrecType cost, shape ~ PrimFnShape prim, m ~ EvaluationMonad prec) => prim -> FailProb prec -> shape ([Value size] -> m [Value size]) -> cost