Traq.Primitives.Class.TypeCheck

Documentation

class (size ~ SizeType prim, ValidPrimShape (PrimFnShape prim)) => TypeCheckPrim prim size | prim -> size where

Type check a primitive given the types of its function arguments, and infer the return types. The typechecker internally checks that the bound arguments are correct, and only gives the user the final type of the partial function.

Minimal complete definition

Nothing

Methods

inferRetTypesPrim :: forall ext' shape m. (m ~ TypeChecker ext', size ~ SizeType ext', shape ~ PrimFnShape prim) => prim -> shape (FnType size) -> m [VarType size]

default inferRetTypesPrim :: forall ext' shape m. (Generic prim, GTypeCheckPrim (Rep prim) size, m ~ TypeChecker ext', size ~ SizeType ext', shape ~ PrimFnShape prim) => prim -> shape (FnType size) -> m [VarType size]