Traq.Primitives.Search.Prelude

Generic Search-like primitives

class HasPrimAny primT where

Primitive any that returns True if there is a solution to the predicate.

Methods

mkAny :: Ident -> primT

build a call to any

getPredicateOfAny :: primT -> Ident

extract the predicate.

class HasPrimSearch primT where

Primitive search that returns a solution to the predicate if exists, otherwise returns a random value. Also returns a bit indicating if a solution was found.

Methods

mkSearch :: Ident -> primT

build a call to search

getPredicateOfSearch :: primT -> Ident

extract the predicate.

Boilerplate functions

Parsing

parsePrimWithPredicates :: String -> Int -> TokenParser () -> Parser [String]

Parse a primitive with a list of predicate names. Example: parsePrimWithPredicates "any" 1 can be used to implement the parser for any.

parsePrimAny :: HasPrimAny primT => String -> TokenParser () -> Parser primT

parsePrimSearch :: HasPrimSearch primT => String -> TokenParser () -> Parser primT

Typecheck

typeCheckSearchPredicate

Arguments

:: TypeCheckable sizeT

=> Ident
name of the predicate function

-> [Ident]
arguments

-> TypeChecker primsT sizeT [VarType sizeT]

typeCheckPrimAny

Arguments

:: (HasPrimAny primT, TypeCheckable sizeT)

=> primT
name of the predicate function

-> [Ident]
arguments

-> TypeChecker primsT sizeT [VarType sizeT]

typeCheckPrimSearch

Arguments

:: (HasPrimAny primT, TypeCheckable sizeT)

=> primT
name of the predicate function

-> [Ident]
arguments

-> TypeChecker primsT sizeT [VarType sizeT]

Evaluation

runSearchPredicateOnAllInputs :: EvaluatablePrimitive primsT primsT => Ident -> [Value SizeT] -> Evaluator primsT SizeT SearchResults

Run the predicate on each input, and return the input along with the result.

hasSolution :: SearchResults -> Bool

countSolutions :: SearchResults -> Int

getSearchOutputs :: SearchResults -> [[Value SizeT]]

evaluatePrimAny :: (HasPrimAny primT, EvaluatablePrimitive primsT primsT) => primT -> [Value SizeT] -> Evaluator primsT SizeT [Value SizeT]

evaluatePrimSearch :: (HasPrimSearch primT, EvaluatablePrimitive primsT primsT) => primT -> [Value SizeT] -> Evaluator primsT SizeT [Value SizeT]

evaluatePrimCount :: EvaluatablePrimitive primsT primsT => Ident -> [Value SizeT] -> Evaluator primsT SizeT [Value SizeT]