Traq.ProtoLang.Cost

Unitary Cost

unitaryQueryCost

Arguments

:: (Num sizeT, Floating costT, UnitaryCostablePrimitive primsT primsT sizeT costT)

=> PrecisionSplittingStrategy

-> costT
precision (l2-norm)

-> Program primsT sizeT
program P

-> OracleTicks costT
oracle ticks

-> costT

unitaryQueryCostE

Arguments

:: (Num sizeT, Floating costT, UnitaryCostablePrimitive primsT primsT sizeT costT, m ~ UnitaryCostCalculator primsT sizeT costT)

=> costT
precision

-> Expr primsT sizeT
expression E

-> m costT

Evaluate the query cost of an expression

unitaryQueryCostS

Arguments

:: (Num sizeT, Floating costT, UnitaryCostablePrimitive primsT primsT sizeT costT, m ~ UnitaryCostCalculator primsT sizeT costT)

=> costT
precision (l2-norm)

-> Stmt primsT sizeT
statement S

-> m costT

Quantum Cost

Expected cost

quantumQueryCost

Arguments

:: (Floating costT, QuantumCostablePrimitive primsT primsT SizeT costT)

=> PrecisionSplittingStrategy

-> costT
failure probability \( \varepsilon \)

-> Program primsT SizeT
program P

-> OracleTicks costT
unitary ticks

-> OracleTicks costT
classical ticks

-> FunInterpCtx SizeT
data injections

-> ProgramState SizeT
state \( \sigma \)

-> costT

quantumQueryCostE

Arguments

:: (Floating costT, QuantumCostablePrimitive primsT primsT SizeT costT, m ~ QuantumCostCalculator primsT SizeT costT)

=> costT
failure probability \( \varepsilon \)

-> ProgramState SizeT
state \( \sigma \)

-> Expr primsT SizeT
statement S

-> m costT

quantumQueryCostS

Arguments

:: (Floating costT, QuantumCostablePrimitive primsT primsT SizeT costT, m ~ QuantumCostCalculator primsT SizeT costT)

=> costT
failure probability \( \varepsilon \)

-> ProgramState SizeT
state \( \sigma \)

-> Stmt primsT SizeT
statement S

-> QuantumCostCalculator primsT SizeT costT costT

Quantum Worst case Cost

quantumMaxQueryCost

Arguments

:: (Num sizeT, Ord costT, Floating costT, QuantumMaxCostablePrimitive primsT primsT sizeT costT)

=> PrecisionSplittingStrategy

-> costT
failure probability eps

-> Program primsT sizeT
program P

-> OracleTicks costT
unitary ticks

-> OracleTicks costT
classical ticks

-> costT

quantumMaxQueryCostE

Arguments

:: (Num sizeT, Ord costT, Floating costT, QuantumMaxCostablePrimitive primsT primsT sizeT costT)

=> costT
failure probability \( \varepsilon \)

-> Expr primsT sizeT
statement S

-> QuantumMaxCostCalculator primsT sizeT costT costT

quantumMaxQueryCostS

Arguments

:: (Num sizeT, Ord costT, Floating costT, QuantumMaxCostablePrimitive primsT primsT sizeT costT)

=> costT
failure probability \( \varepsilon \)

-> Stmt primsT sizeT
statement S

-> QuantumMaxCostCalculator primsT sizeT costT costT

Bound on runtime

quantumQueryCostBound

Arguments

:: (Ord costT, Floating costT, QuantumCostablePrimitive primsT primsT SizeT costT)

=> PrecisionSplittingStrategy

-> costT
failure probability \( \varepsilon \)

-> Program primsT SizeT
program P

-> OracleTicks costT
unitary ticks

-> OracleTicks costT
classical ticks

-> FunInterpCtx SizeT
data injections

-> ProgramState SizeT
state \( \sigma \)

-> costT

The bound on the true expected runtime which fails with probability <= eps.

Types and Monad

data UnitaryCostEnv primT sizeT costT

Environment to compute the unitary cost

data QuantumMaxCostEnv primT sizeT costT

Environment to compute the worst case quantum cost

data QuantumCostEnv primsT sizeT costT

Data Types

type OracleTicks costT = Map Ident costT

Tick constant for each oracle.

Lenses

_classicalTicks :: (HasClassicalTicks p, costT ~ CostType p) => Lens' p (OracleTicks costT)

_unitaryTicks :: (HasUnitaryTicks p, costT ~ CostType p) => Lens' p (OracleTicks costT)

_funInterpCtx :: (HasFunInterpCtx p, sizeT ~ SizeType p) => Lens' p (FunInterpCtx sizeT)

_unitaryCostEnv :: (HasUnitaryCostEnv p, primT ~ PrimitiveType p, sizeT ~ SizeType p, costT ~ CostType p) => Lens' p (UnitaryCostEnv primT sizeT costT)

_quantumMaxCostEnv :: (HasQuantumMaxCostEnv p, primT ~ PrimitiveType p, sizeT ~ SizeType p, costT ~ CostType p) => Lens' p (QuantumMaxCostEnv primT sizeT costT)

Cost Monads

type UnitaryCostCalculator primsT sizeT costT = ReaderT (UnitaryCostEnv primsT sizeT costT) Maybe

Monad to compute unitary cost.

class (TypeCheckablePrimitive primT sizeT, Show costT) => UnitaryCostablePrimitive primsT primT sizeT costT where

Primitives that have a unitary cost

Methods

unitaryQueryCostPrimitive :: costT -> primT -> [Ident] -> UnitaryCostCalculator primsT sizeT costT costT

type QuantumMaxCostCalculator primsT sizeT costT = ReaderT (QuantumMaxCostEnv primsT sizeT costT) Maybe

class UnitaryCostablePrimitive primsT primT sizeT costT => QuantumMaxCostablePrimitive primsT primT sizeT costT where

Primitives that have a quantum max cost

Methods

quantumMaxQueryCostPrimitive :: costT -> primT -> QuantumMaxCostCalculator primsT sizeT costT costT

type QuantumCostCalculator primsT sizeT costT = ReaderT (QuantumCostEnv primsT sizeT costT) Maybe

class (QuantumMaxCostablePrimitive primsT primT sizeT costT, EvaluatablePrimitive primsT primT) => QuantumCostablePrimitive primsT primT sizeT costT where

Primitives that have a input dependent expected quantum cost

Methods

quantumQueryCostPrimitive :: costT -> primT -> [Value sizeT] -> QuantumCostCalculator primsT sizeT costT costT

Precision Splitting Strategies

data PrecisionSplittingStrategy

Constructors

SplitSimple

SplitUsingNeedsEps

class HasPrecisionSplittingStrategy p where

Methods

_precSplitStrat :: Lens' p PrecisionSplittingStrategy

class HasNeedsEps p where

Predicate for checking if a exprstmtprog can fail

Methods

needsEps :: (primT ~ PrimitiveType p, sizeT ~ SizeType p, MonadReader env m, HasFunCtx env, sizeT ~ SizeType env, primT ~ PrimitiveType env) => p -> m Bool

splitEps :: forall primT sizeT costT env m. (Floating costT, Monad m, MonadReader env m, HasFunCtx env, HasPrecisionSplittingStrategy env, HasNeedsEps (Stmt primT sizeT), primT ~ PrimitiveType env, sizeT ~ SizeType env) => costT -> [Stmt primT sizeT] -> m [costT]