Traq.ProtoLang.Eval

Evaluating Basic Expressions

type ProgramState sizeT = Context (Value sizeT)

The deterministic state of the program

class HasProgramState p where

Methods

_state :: sizeT ~ SizeType p => Lens' p (ProgramState sizeT)

evalBasicExpr :: (MonadReader env m, HasProgramState env, sizeT ~ SizeType env, sizeT ~ SizeT) => BasicExpr sizeT -> m (Value sizeT)

Evaluations

runProgram :: forall ext precT m. (Evaluatable ext SizeT precT, EvalReqs SizeT precT, m ~ EvaluationMonad precT) => Program ext -> FunInterpCtx SizeT -> [Value SizeT] -> m [Value SizeT]

Entry-point: run the program (i.e. last function)

Values

defaultV :: forall sizeT. sizeT ~ SizeT => VarType sizeT -> Value sizeT

default value of a given type

validateValueType :: forall sizeT. sizeT ~ SizeT => VarType sizeT -> Value sizeT -> Bool

Check if a given runtime value is of a given type.

class CoerceValue t where

Coerce a haskell value into our value type.

Methods

safeToValue :: t -> Maybe (Value SizeT)

safely convert a runtime value to Value.

safeFromValue :: Value SizeT -> Maybe t

safely convert a Value to a runtime value.

toValue :: CoerceValue t => t -> Value SizeT

Convert an input into Value

fromValue :: CoerceValue t => Value SizeT -> t

Convert a Value to the given type.

valueToBool :: Value SizeT -> Bool

domainSize :: Integral sizeT => VarType sizeT -> sizeT

Size of the value set of a given type.

bitsize :: Num size => VarType size -> Maybe size

Bitsize

domain :: Integral sizeT => VarType sizeT -> [Value sizeT]

Set of all values of a given type

Types and Monad

type EvaluationMonad precT = ExpMonad precT

Base probability monad to evaluate the program.

class (EvalReqs sizeT precT, sizeT ~ SizeType ext, precT ~ PrecType ext) => Evaluatable ext sizeT precT | ext -> sizeT precT where

Primitives that support evaluation.

Minimal complete definition

Nothing

Methods

eval :: forall ext' m. (Evaluatable ext' sizeT precT, m ~ Evaluator ext', SizeType ext' ~ sizeT, PrecType ext' ~ precT) => ext -> ProgramState sizeT -> m [Value sizeT]

default eval :: forall ext' m. (Generic ext, GEvaluatable (Rep ext) sizeT precT, Evaluatable ext' sizeT precT, m ~ Evaluator ext', SizeType ext' ~ sizeT, PrecType ext' ~ precT) => ext -> ProgramState sizeT -> m [Value sizeT]

type EvalReqs sizeT precT = (sizeT ~ SizeT, ProbType precT, RVType precT precT, Fractional precT)

Constraints to be satisfied to support evaluation.

type family EvalArgs f ext

eval1 :: forall ext size prec m. (Eval1 f, Evaluatable ext size prec, m ~ Evaluator ext) => f ext -> EvalArgs f ext -> m (EvalRets f ext)

Evaluation

type FunInterp sizeT = [Value sizeT] -> [Value sizeT]

Inject runtime data into a program

type FunInterpCtx sizeT = Map Ident (FunInterp sizeT)

A mapping of data intepretations

class HasFunInterpCtx p where

Methods

_funInterpCtx :: sizeT ~ SizeType p => Lens' p (FunInterpCtx sizeT)

data EvaluationEnv ext

Environment for evaluation

class HasEvaluationEnv p ext | p -> ext where

Methods

_evaluationEnv :: Lens' p (EvaluationEnv ext)

type Evaluator ext = ReaderT (EvaluationEnv ext) (EvaluationMonad (PrecType ext))

Non-deterministic evaluation monad (i.e. no state)