Traq.ProtoLang.Syntax

Syntax

data MetaParam sizeT

Compile-time constant parameters

Constructors

MetaName String

MetaSize sizeT

MetaValue Integer

Basic Types

newtype VarType sizeT

Types

Constructors

Fin sizeT

_Fin :: Traversal' (VarType sizeT) sizeT

data Value sizeT

Basic Values

Constructors

FinV sizeT
value of type Fin n

Basic Operations

data UnOp

Unary operations

Constructors

NotOp

data BinOp

Binary operations

Constructors

AddOp

LEqOp

GtOp

AndOp

data NAryOp

Operations which take multiple arguments

Constructors

MultiOrOp

data BasicExpr sizeT

Basic arithmetic and logical expressions

Constructors

VarE

Fields

var :: Ident

ParamE

Fields

param :: Ident

ConstE

Fields

val :: Value sizeT

ty :: VarType sizeT

UnOpE

Fields

un_op :: UnOp

operand :: BasicExpr sizeT

BinOpE

Fields

bin_op :: BinOp

lhs, rhs :: BasicExpr sizeT

TernaryE

Fields

branch, lhs, rhs :: BasicExpr sizeT

NAryE

Fields

op :: NAryOp

operands :: [BasicExpr sizeT]

(.<=.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

notE :: BasicExpr sizeT -> BasicExpr sizeT

(.+.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

(.&&.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

Calls and Primitives

data FunctionCallKind primT

Either call an existing function, or an existing primitive.

Constructors

FunctionCall Ident

PrimitiveCall

Fields

prim :: primT

Expressions and Statements

data Expr primT sizeT

An expression in the prototype language. It appears as the RHS of an assignment statement.

Constructors

BasicExprE

Fields

basic_expr :: BasicExpr sizeT

FunCallE

Fields

fun_kind :: FunctionCallKind primT

args :: [Ident]

data Stmt primT sizeT

A statement in the prototype language.

Constructors

ExprS

Fields

rets :: [Ident]

expr :: Expr primT sizeT

IfThenElseS

Fields

cond :: Ident

s_true, s_false :: Stmt primT sizeT

SeqS [Stmt primT sizeT]

data FunBody primT sizeT

The body of a function.

Constructors

FunBody

Fields

param_names, ret_names :: [Ident]

body_stmt :: Stmt primT sizeT

data FunDef primT sizeT

A function definition or declaration in the prototype language.

Constructors

FunDef

Fields

param_types, ret_types :: [VarType sizeT]

mbody :: Maybe (FunBody primT sizeT)

data NamedFunDef primT sizeT

A function with a name

Constructors

NamedFunDef

Fields

fun_name :: Ident

fun_def :: FunDef primT sizeT

type FunCtx primT sizeT = Context (FunDef primT sizeT)

A function context contains a list of functions

data Program primT sizeT

A program is a function context with a statement (which acts like the main)

Constructors

Program

Fields

funCtx :: FunCtx primT sizeT

stmt :: Stmt primT sizeT

Lenses

class HasFunCtx p where

Methods

_funCtx :: (primT ~ PrimitiveType p, sizeT ~ SizeType p) => Lens' p (FunCtx primT sizeT)