Traq.ProtoLang.TypeCheck

Typing Context

type TypingCtx sizeT = Context (VarType sizeT)

A context mapping variables to their types.

class HasTypingCtx p where

Methods

_typingCtx :: sizeT ~ SizeType p => Lens' p (TypingCtx sizeT)

Typeable size types

class (Eq sizeT, Show sizeT, Num sizeT) => TypeCheckable sizeT where

SizeT that can be type-checked. Needs to have a boolean type, and max of two types.

Methods

tbool :: VarType sizeT

tmax :: VarType sizeT -> VarType sizeT -> VarType sizeT

irange :: sizeT -> Int

Types

type TypingEnv primT sizeT = FunCtx primT sizeT

Environment for type checking

type TypeChecker primT sizeT = ReaderT (TypingEnv primT sizeT) (StateT (TypingCtx sizeT) (Either String))

The TypeChecker monad

class TypeCheckablePrimitive primT sizeT where

Methods

typeCheckPrimitive

Arguments

:: TypeCheckable sizeT

=> primT
primitive

-> [Ident]
arguments

-> TypeChecker primsT sizeT [VarType sizeT]

Checkers

typeCheckBasicExpr :: forall sizeT m. (TypeCheckable sizeT, MonadError String m, MonadReader (TypingCtx sizeT) m) => BasicExpr sizeT -> m (VarType sizeT)

typeCheckStmt :: forall primT sizeT. (TypeCheckablePrimitive primT sizeT, TypeCheckable sizeT) => Stmt primT sizeT -> TypeChecker primT sizeT ()

Typecheck a statement, given the current context and function definitions. If successful, the typing context is updated.

typeCheckFun :: (TypeCheckablePrimitive primT sizeT, TypeCheckable sizeT) => FunCtx primT sizeT -> FunDef primT sizeT -> Either String ()

Type check a single function.

typeCheckProg :: (TypeCheckablePrimitive primT sizeT, TypeCheckable sizeT) => TypingCtx sizeT -> Program primT sizeT -> Either String ()

Type check a full program (i.e. list of functions).