Traq.Data.Symbolic

Description

Based on https://hackage.haskell.org/package/numbers-3000.2.0.2

Documentation

data Sym a

Symbolic numbers over some base type for the literals.

var :: String -> Sym a

Create a variable.

con :: a -> Sym a

Create a constant (useful when it is not a literal).

subst :: (Num a, Eq a) => String -> Sym a -> Sym a -> Sym a

The expression subst x v e substitutes the expression v for each occurence of the variable x in e.

unSym :: Show a => Sym a -> a

simpl :: (Num a, Eq a) => Sym a -> Sym a

Simplify by re-assembling the expression