Traq.ProtoLang.Syntax

Syntax

data MetaParam sizeT

Compile-time constant parameters

Constructors

MetaName String

MetaSize sizeT

MetaValue Integer

Basic Types

data VarType sizeT

Types

Constructors

Fin sizeT

Bitvec sizeT

Arr sizeT (VarType sizeT)

Tup [VarType sizeT]

_Fin :: Traversal' (VarType sizeT) sizeT

_Arr :: Traversal' (VarType sizeT) (sizeT, VarType sizeT)

_Tup :: Traversal' (VarType sizeT) [VarType sizeT]

data Value sizeT

Basic Values

Constructors

FinV sizeT
value of type Fin n

ArrV [Value sizeT]
value of type Arr n t

TupV [Value sizeT]
tuple value

_FinV :: Traversal' (Value sizeT) sizeT

_ArrV :: Traversal' (Value sizeT) [Value sizeT]

_TupV :: Traversal' (Value sizeT) [Value sizeT]

Function types

data FnType size

Constructors

FnType [VarType size] [VarType size]

Basic Operations

data UnOp

Unary operations

Constructors

NotOp

AnyOp

MajOp

data BinOp

Binary operations

Constructors

AddOp

MulOp

SubOp

XorOp

LEqOp

LtOp

AndOp

EqOp

VecSelectOp

data NAryOp

Operations which take multiple arguments

Constructors

MultiOrOp

data BasicExpr sizeT

Basic arithmetic and logical expressions

Constructors

VarE

Fields

var :: Ident

ParamE

Fields

param :: Ident

DefaultE

Fields

ty :: VarType sizeT

ConstE

Fields

val :: Value sizeT

ty :: VarType sizeT

UnOpE

Fields

un_op :: UnOp

operand :: BasicExpr sizeT

BinOpE

Fields

bin_op :: BinOp

lhs, rhs :: BasicExpr sizeT

TernaryE

Fields

branch, lhs, rhs :: BasicExpr sizeT

NAryE

Fields

op :: NAryOp

operands :: [BasicExpr sizeT]

IndexE

Fields

arr_expr :: BasicExpr sizeT

ix_val :: sizeT

DynIndexE

Fields

arr_expr, ix_expr :: BasicExpr sizeT

UpdateArrE

Fields

arr_expr, ix_expr, rhs :: BasicExpr sizeT

ProjectE

Fields

tup_expr :: BasicExpr sizeT

tup_ix_val :: Int

(.<=.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

notE :: BasicExpr sizeT -> BasicExpr sizeT

(.+.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

(.&&.) :: BasicExpr sizeT -> BasicExpr sizeT -> BasicExpr sizeT

Expressions and Statements

data DistrExpr sizeT

An expression denoting a probablity distribution.

Constructors

UniformE

Fields

sample_ty :: VarType sizeT

BernoulliE

Fields

prob_one :: Double

data Expr ext

An expression in the prototype language. It appears as the RHS of an assignment statement.

Constructors

BasicExprE

Fields

basic_expr :: BasicExpr (SizeType ext)

RandomSampleE

Fields

distr_expr :: DistrExpr (SizeType ext)

FunCallE

Fields

fname :: Ident

args :: [Ident]

PrimCallE

Fields

prim :: ext

LoopE

Fields

initial_args :: [Ident]

loop_body_fun :: Ident

data Stmt ext

A statement in the prototype language.

Constructors

ExprS

Fields

rets :: [Ident]

expr :: Expr ext

IfThenElseS

Fields

cond :: Ident

s_true, s_false :: Stmt ext

SeqS [Stmt ext]

data FunBody ext

The body of a function.

Constructors

FunBody

Fields

param_names, ret_names :: [Ident]

body_stmt :: Stmt ext

data FunDef ext

A function definition or declaration in the prototype language.

Constructors

FunDef

Fields

param_types, ret_types :: [VarType (SizeType ext)]

mbody :: Maybe (FunBody ext)

data NamedFunDef ext

A function with a name

Constructors

NamedFunDef

Fields

fun_name :: Ident

fun_def :: FunDef ext

newtype Program ext

A program is a list of named functions, with the last being the entry point.

Constructors

Program [NamedFunDef ext]

FunCtx

type FunCtx ext = Context (FunDef ext)

A function context contains a list of functions

namedFunsToFunCtx :: Foldable f => f (NamedFunDef ext) -> FunCtx ext

programToFunCtx :: Program ext -> FunCtx ext

funCtxToNamedFuns :: FunCtx ext -> [NamedFunDef ext]

Core

data Core sizeT precT

Void extension (i.e. only use the core language) Usage: p :: Program (Core sizeT precT)

Core Language with default types.

type Core' = Core SizeT Double

Simple void extension with integer size and double prec. Usage: p :: Program Core'

Lenses

class HasFunCtx p ext | p -> ext where

Methods

_funCtx :: Lens' p (FunCtx ext)