Traq.Utils.Printing

Documentation

class ToCodeString a where

Methods

build :: (MonadWriter [String] m, MonadFail m) => a -> m ()

toCodeString :: ToCodeString a => a -> String

toCodeWord :: ToCodeString a => a -> String

toCodeStringM :: (ToCodeString a, MonadFail m) => a -> m String

Simple builders

commaList :: [String] -> String

wrapNonEmpty :: String -> String -> String -> String

Wrap a non-empty string by the delimiters, and pass empty strings through as-is.

Basic writers

putWord :: MonadWriter [w] m => w -> m ()

putLine :: MonadWriter [w] m => w -> m ()

Add a line of code.

putComment :: MonadWriter [String] m => String -> m ()

endl :: MonadWriter [String] m => m ()

fromBuild :: (ToCodeString a, MonadWriter [String] m, MonadFail m) => a -> m String

listenWord :: (MonadWriter [String] m, MonadFail m) => m () -> m String

Generic writer transformers

mapped :: MonadWriter [w] m => (w -> w) -> m a -> m a

Map each line in a block of code.

joined :: MonadWriter [w] m => ([w] -> w) -> m a -> m a

Join a block of code using a joining function.

delimitedBlock :: MonadWriter [String] m => String -> String -> m a -> m a

Specific writer transformers

indented :: MonadWriter [String] m => m a -> m a

Indent the block of code.

concatenated :: MonadWriter [[w]] m => m a -> m a

Concatenate the block of code into a single line.

unworded :: MonadWriter [String] m => m a -> m a

unwords the lines in a block of code.

unlined :: MonadWriter [String] m => m a -> m a

unlines the lines in a block of code.

commented :: MonadWriter [String] m => m a -> m a

bracedBlock :: MonadWriter [String] m => m a -> m a

bracedBlockWith :: MonadWriter [String] m => String -> m a -> m a

doEndBlock :: MonadWriter [String] m => m a -> m a

prepended :: MonadWriter [String] m => String -> m a -> m a

Prepend the string to the first line out the output